backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Backend for the Todo Application with AI Agent and MCP Server Integration"
authors = [
    { name = "Claude Code", email = "claude@example.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlmodel>=0.0.22",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
    "openai>=1.55.0",
    "pydantic>=2.9.0",
    "openai-agents>=0.6.9",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
